refactor(api): project_configs 자산 라우트를 skills·agents·mcp 모듈로 추출
22개 라우트 이동 (skills 7 · agents 7 · mcp 8). 핸들러 본문·경로·함수명 무변경.

**추출을 라인 범위가 아니라 이름 기반으로 했다.** AST 가 계산한 (데코레이터
시작, end_lineno)로 잘라내므로 오프바이원이 발생할 수 없다 — claude_sessions
에서 손으로 고른 범위가 1줄 짧아 return 문을 잃은 사고의 재발 방지다.

R2b(split_audit) 통과: 원본 61개 정의(60 핸들러 + _get_db_filtered_projects)
전부 존재·바이트 동일, 중복 0건. 라우트 테이블 60개 동일.

세 모듈 모두 3세그먼트 이상이라 include 순서 제약이 없다. __init__.py 에
전역→프로젝트 순서 계약을 주석으로 명시했다.
