+++ src/backend/api/project_configs/_legacy.py
@@ -13,28 +13,17 @@ from fastapi.responses import StreamingResponse
 
 from api.deps import get_current_user_optional
 from models.project_config import (
-    AgentConfig,
-    AgentContentResponse,
-    AgentCreateRequest,
-    AgentUpdateRequest,
     CommandConfig,
     CommandContentResponse,
     CommandCreateRequest,
     CommandUpdateRequest,
-    CopyAgentRequest,
     CopyCommandRequest,
     CopyHookRequest,
-    CopyMCPRequest,
     CopyRuleRequest,
-    CopySkillRequest,
     ExternalPathRequest,
     GlobalConfigSummary,
     HookEntryRequest,
     HooksUpdateRequest,
-    MCPServerConfig,
-    MCPServerCreateRequest,
-    MCPServerUpdateRequest,
-    MCPToggleRequest,
     MemoryConfig,
     MemoryContentResponse,
     MemoryCreateRequest,
@@ -45,10 +34,6 @@ from models.project_config import (
     RuleContentResponse,
     RuleCreateRequest,
     RuleUpdateRequest,
-    SkillConfig,
-    SkillContentResponse,
-    SkillCreateRequest,
-    SkillUpdateRequest,
 )
 from services.audit_service import AuditAction, audit_config_change
 from services.project_config_monitor import get_project_config_monitor
@@ -575,506 +560,14 @@ async def get_project_summary(project_id: str) -> ProjectConfigSummary:
 # ========================================
 
 
-@router.get("/skills/all", response_model=list[SkillConfig])
-async def list_all_skills() -> list[SkillConfig]:
-    """Get all skills from all monitored projects.
-
-    Returns:
-        List of all skills across all projects
-    """
-    monitor = get_project_config_monitor()
-    return monitor.get_all_skills()
-
-
-@router.get("/{project_id}/skills", response_model=list[SkillConfig])
-async def list_project_skills(project_id: str) -> list[SkillConfig]:
-    """Get all skills for a specific project.
-
-    Args:
-        project_id: Project identifier
-
-    Returns:
-        List of skills for the project
-    """
-    monitor = get_project_config_monitor()
-    skills = monitor.get_project_skills(project_id)
-
-    if not skills:
-        # Check if project exists
-        summary = monitor.get_project_summary(project_id)
-        if summary is None:
-            raise HTTPException(status_code=404, detail=f"Project not found: {project_id}")
-
-    return skills
-
-
-@router.get("/{project_id}/skills/{skill_id}/content", response_model=SkillContentResponse)
-async def get_skill_content(project_id: str, skill_id: str) -> SkillContentResponse:
-    """Get full content of a skill including SKILL.md and references.
-
-    Args:
-        project_id: Project identifier
-        skill_id: Skill identifier (directory name)
-
-    Returns:
-        Skill configuration with full content and reference paths
-    """
-    monitor = get_project_config_monitor()
-    skill, content, references = monitor.get_skill_content(project_id, skill_id)
-
-    if skill is None:
-        raise HTTPException(
-            status_code=404,
-            detail=f"Skill not found: {skill_id} in project {project_id}",
-        )
-
-    return SkillContentResponse(
-        skill=skill,
-        content=content,
-        references=references,
-    )
-
-
-@router.post("/{project_id}/skills", response_model=SkillConfig)
-async def create_skill(project_id: str, request: SkillCreateRequest) -> SkillConfig:
-    """Create a new skill.
-
-    Args:
-        project_id: Project identifier
-        request: Skill create request
-
-    Returns:
-        Created skill configuration
-    """
-    monitor = get_project_config_monitor()
-    result = monitor.create_skill(project_id, request.skill_id, request.content)
-
-    if result is None:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to create skill: {request.skill_id}. Check if project exists and skill ID is unique.",
-        )
-
-    audit_config_change(AuditAction.CONFIG_CREATED, "skill", request.skill_id, project_id)
-    return result
-
-
-@router.put("/{project_id}/skills/{skill_id}")
-async def update_skill(project_id: str, skill_id: str, request: SkillUpdateRequest) -> dict:
-    """Update skill content.
-
-    Args:
-        project_id: Project identifier
-        skill_id: Skill identifier
-        request: Update request with new content
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.update_skill_content(project_id, skill_id, request.content):
-        audit_config_change(AuditAction.CONFIG_UPDATED, "skill", skill_id, project_id)
-        return {
-            "success": True,
-            "message": f"Updated skill: {skill_id}",
-            "project_id": project_id,
-            "skill_id": skill_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to update skill: {skill_id}. Check if project and skill exist.",
-        )
-
-
-@router.delete("/{project_id}/skills/{skill_id}")
-async def delete_skill(project_id: str, skill_id: str) -> dict:
-    """Delete a skill.
-
-    Args:
-        project_id: Project identifier
-        skill_id: Skill identifier
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.delete_skill(project_id, skill_id):
-        audit_config_change(AuditAction.CONFIG_DELETED, "skill", skill_id, project_id)
-        return {
-            "success": True,
-            "message": f"Deleted skill: {skill_id}",
-            "project_id": project_id,
-            "skill_id": skill_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to delete skill: {skill_id}. Check if project and skill exist.",
-        )
-
-
 # ========================================
 # Agents
 # ========================================
 
 
-@router.get("/agents/all", response_model=list[AgentConfig])
-async def list_all_agents() -> list[AgentConfig]:
-    """Get all agents from all monitored projects.
-
-    Returns:
-        List of all agents across all projects
-    """
-    monitor = get_project_config_monitor()
-    return monitor.get_all_agents()
-
-
-@router.get("/{project_id}/agents", response_model=list[AgentConfig])
-async def list_project_agents(project_id: str) -> list[AgentConfig]:
-    """Get all agents for a specific project.
-
-    Args:
-        project_id: Project identifier
-
-    Returns:
-        List of agents for the project
-    """
-    monitor = get_project_config_monitor()
-    agents = monitor.get_project_agents(project_id)
-
-    if not agents:
-        # Check if project exists
-        summary = monitor.get_project_summary(project_id)
-        if summary is None:
-            raise HTTPException(status_code=404, detail=f"Project not found: {project_id}")
-
-    return agents
-
-
-@router.get("/{project_id}/agents/{agent_id}/content", response_model=AgentContentResponse)
-async def get_agent_content(project_id: str, agent_id: str) -> AgentContentResponse:
-    """Get full content of an agent.
-
-    Args:
-        project_id: Project identifier
-        agent_id: Agent identifier
-
-    Returns:
-        Agent configuration with full content
-    """
-    monitor = get_project_config_monitor()
-    agent, content = monitor.get_agent_content(project_id, agent_id)
-
-    if agent is None:
-        raise HTTPException(
-            status_code=404,
-            detail=f"Agent not found: {agent_id} in project {project_id}",
-        )
-
-    return AgentContentResponse(agent=agent, content=content)
-
-
-@router.post("/{project_id}/agents", response_model=AgentConfig)
-async def create_agent(project_id: str, request: AgentCreateRequest) -> AgentConfig:
-    """Create a new agent.
-
-    Args:
-        project_id: Project identifier
-        request: Agent create request
-
-    Returns:
-        Created agent configuration
-    """
-    monitor = get_project_config_monitor()
-    result = monitor.create_agent(project_id, request.agent_id, request.content, request.is_shared)
-
-    if result is None:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to create agent: {request.agent_id}. Check if project exists and agent ID is unique.",
-        )
-
-    audit_config_change(AuditAction.CONFIG_CREATED, "agent", request.agent_id, project_id)
-    return result
-
-
-@router.put("/{project_id}/agents/{agent_id}")
-async def update_agent(project_id: str, agent_id: str, request: AgentUpdateRequest) -> dict:
-    """Update agent content.
-
-    Args:
-        project_id: Project identifier
-        agent_id: Agent identifier
-        request: Update request with new content
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.update_agent_content(project_id, agent_id, request.content):
-        audit_config_change(AuditAction.CONFIG_UPDATED, "agent", agent_id, project_id)
-        return {
-            "success": True,
-            "message": f"Updated agent: {agent_id}",
-            "project_id": project_id,
-            "agent_id": agent_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to update agent: {agent_id}. Check if project and agent exist.",
-        )
-
-
-@router.delete("/{project_id}/agents/{agent_id}")
-async def delete_agent(project_id: str, agent_id: str) -> dict:
-    """Delete an agent.
-
-    Args:
-        project_id: Project identifier
-        agent_id: Agent identifier
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.delete_agent(project_id, agent_id):
-        audit_config_change(AuditAction.CONFIG_DELETED, "agent", agent_id, project_id)
-        return {
-            "success": True,
-            "message": f"Deleted agent: {agent_id}",
-            "project_id": project_id,
-            "agent_id": agent_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to delete agent: {agent_id}. Check if project and agent exist.",
-        )
-
-
 # ========================================
 # MCP Servers
 # ========================================
-
-
-@router.get("/{project_id}/mcp", response_model=list[MCPServerConfig])
-async def list_project_mcp_servers(project_id: str) -> list[MCPServerConfig]:
-    """Get MCP server configuration for a project.
-
-    Args:
-        project_id: Project identifier
-
-    Returns:
-        List of MCP server configurations
-    """
-    monitor = get_project_config_monitor()
-    servers = monitor.get_project_mcp_config(project_id)
-
-    if not servers:
-        # Check if project exists
-        summary = monitor.get_project_summary(project_id)
-        if summary is None:
-            raise HTTPException(status_code=404, detail=f"Project not found: {project_id}")
-
-    return servers
-
-
-@router.post("/{project_id}/mcp/{server_id}/enable")
-async def enable_mcp_server(project_id: str, server_id: str) -> dict:
-    """Enable an MCP server in a project.
-
-    Updates the mcp.json file to set disabled: false.
-
-    Args:
-        project_id: Project identifier
-        server_id: MCP server identifier
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.enable_mcp_server(project_id, server_id):
-        return {
-            "success": True,
-            "message": f"Enabled MCP server: {server_id}",
-            "project_id": project_id,
-            "server_id": server_id,
-            "disabled": False,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to enable MCP server: {server_id}. Check if project and server exist.",
-        )
-
-
-@router.post("/{project_id}/mcp/{server_id}/disable")
-async def disable_mcp_server(project_id: str, server_id: str) -> dict:
-    """Disable an MCP server in a project.
-
-    Updates the mcp.json file to set disabled: true.
-
-    Args:
-        project_id: Project identifier
-        server_id: MCP server identifier
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.disable_mcp_server(project_id, server_id):
-        return {
-            "success": True,
-            "message": f"Disabled MCP server: {server_id}",
-            "project_id": project_id,
-            "server_id": server_id,
-            "disabled": True,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to disable MCP server: {server_id}. Check if project and server exist.",
-        )
-
-
-@router.post("/{project_id}/mcp/{server_id}/toggle")
-async def toggle_mcp_server(project_id: str, server_id: str, request: MCPToggleRequest) -> dict:
-    """Toggle MCP server enabled/disabled state.
-
-    Args:
-        project_id: Project identifier
-        server_id: MCP server identifier
-        request: Toggle request with enabled state
-
-    Returns:
-        Success status with new state
-    """
-    monitor = get_project_config_monitor()
-
-    if request.enabled:
-        success = monitor.enable_mcp_server(project_id, server_id)
-    else:
-        success = monitor.disable_mcp_server(project_id, server_id)
-
-    if success:
-        return {
-            "success": True,
-            "message": f"{'Enabled' if request.enabled else 'Disabled'} MCP server: {server_id}",
-            "project_id": project_id,
-            "server_id": server_id,
-            "disabled": not request.enabled,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to toggle MCP server: {server_id}",
-        )
-
-
-@router.put("/{project_id}/mcp/{server_id}", response_model=MCPServerConfig)
-async def update_mcp_server(
-    project_id: str, server_id: str, request: MCPServerUpdateRequest
-) -> MCPServerConfig:
-    """Update an MCP server configuration.
-
-    Args:
-        project_id: Project identifier
-        server_id: MCP server identifier
-        request: Update request with new values
-
-    Returns:
-        Updated MCP server configuration
-    """
-    monitor = get_project_config_monitor()
-
-    result = monitor.update_mcp_server(
-        project_id=project_id,
-        server_id=server_id,
-        command=request.command,
-        args=request.args,
-        env=request.env,
-        disabled=request.disabled,
-        note=request.note,
-    )
-
-    if result is None:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to update MCP server: {server_id}. Check if project and server exist.",
-        )
-
-    audit_config_change(AuditAction.CONFIG_UPDATED, "mcp_server", server_id, project_id)
-    return result
-
-
-@router.post("/{project_id}/mcp", response_model=MCPServerConfig)
-async def create_mcp_server(project_id: str, request: MCPServerCreateRequest) -> MCPServerConfig:
-    """Create a new MCP server configuration.
-
-    Args:
-        project_id: Project identifier
-        request: Create request with server details
-
-    Returns:
-        Created MCP server configuration
-    """
-    monitor = get_project_config_monitor()
-
-    result = monitor.create_mcp_server(
-        project_id=project_id,
-        server_id=request.server_id,
-        command=request.command,
-        args=request.args,
-        env=request.env,
-        disabled=request.disabled,
-        note=request.note,
-    )
-
-    if result is None:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to create MCP server: {request.server_id}. "
-            "Check if project exists and server ID is unique.",
-        )
-
-    audit_config_change(AuditAction.CONFIG_CREATED, "mcp_server", request.server_id, project_id)
-    return result
-
-
-@router.delete("/{project_id}/mcp/{server_id}")
-async def delete_mcp_server(project_id: str, server_id: str) -> dict:
-    """Delete an MCP server configuration.
-
-    Args:
-        project_id: Project identifier
-        server_id: MCP server identifier
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.delete_mcp_server(project_id, server_id):
-        audit_config_change(AuditAction.CONFIG_DELETED, "mcp_server", server_id, project_id)
-        return {
-            "success": True,
-            "message": f"Deleted MCP server: {server_id}",
-            "project_id": project_id,
-            "server_id": server_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to delete MCP server: {server_id}. Check if project and server exist.",
-        )
 
 
 # ========================================
@@ -1659,99 +1152,6 @@ async def copy_rule(project_id: str, rule_id: str, request: CopyRuleRequest) -> 
 # ========================================
 # Copy Operations
 # ========================================
-
-
-@router.post("/{project_id}/skills/{skill_id}/copy")
-async def copy_skill(project_id: str, skill_id: str, request: CopySkillRequest) -> dict:
-    """Copy a skill to another project.
-
-    Args:
-        project_id: Source project identifier
-        skill_id: Skill identifier to copy
-        request: Copy request with target project
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    result = monitor.copy_skill(project_id, skill_id, request.target_project_id)
-
-    if result:
-        return {
-            "success": True,
-            "message": f"Copied skill '{skill_id}' to project '{request.target_project_id}'",
-            "source_project_id": project_id,
-            "target_project_id": request.target_project_id,
-            "skill_id": skill_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to copy skill '{skill_id}'. Check if source and target projects exist.",
-        )
-
-
-@router.post("/{project_id}/agents/{agent_id}/copy")
-async def copy_agent(project_id: str, agent_id: str, request: CopyAgentRequest) -> dict:
-    """Copy an agent to another project.
-
-    Args:
-        project_id: Source project identifier
-        agent_id: Agent identifier to copy
-        request: Copy request with target project
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    result = monitor.copy_agent(project_id, agent_id, request.target_project_id)
-
-    if result:
-        return {
-            "success": True,
-            "message": f"Copied agent '{agent_id}' to project '{request.target_project_id}'",
-            "source_project_id": project_id,
-            "target_project_id": request.target_project_id,
-            "agent_id": agent_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to copy agent '{agent_id}'. Check if source and target projects exist.",
-        )
-
-
-@router.post("/{project_id}/mcp/{server_id}/copy")
-async def copy_mcp_server(project_id: str, server_id: str, request: CopyMCPRequest) -> dict:
-    """Copy an MCP server to another project.
-
-    Args:
-        project_id: Source project identifier
-        server_id: MCP server identifier to copy
-        request: Copy request with target project
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    result = monitor.copy_mcp_server(project_id, server_id, request.target_project_id)
-
-    if result:
-        return {
-            "success": True,
-            "message": f"Copied MCP server '{server_id}' to project '{request.target_project_id}'",
-            "source_project_id": project_id,
-            "target_project_id": request.target_project_id,
-            "server_id": server_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to copy MCP server '{server_id}'. Check if source and target projects exist.",
-        )
 
 
 @router.post("/{project_id}/hooks/{event}/{index}/copy")
